shift right to avoid possibility of int64 overflow in rw_distance calculation
# Conflicts:
#	src/doom/p_setup.c
#	src/doom/r_defs.h
#	src/doom/r_segs.c

+++ src/doom/r_defs.h
@@ -254,7 +254,7 @@ typedef struct
     sector_t*	frontsector;
     sector_t*	backsector;
     
-    int64_t	length; // [crispy] fix long wall wobble
+    uint32_t	length; // [crispy] fix long wall wobble
 } seg_t;
